Remove `include/grpcpp/impl/codegen/rpc_method.h` (#31269)
