+++ src/compiler/cpp_generator.cc
@@ -146,7 +146,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/support/message_allocator.h",
         "grpcpp/support/method_handler.h",
         "grpcpp/impl/codegen/proto_utils.h",
-        "grpcpp/impl/codegen/rpc_method.h",
+        "grpcpp/impl/rpc_method.h",
         "grpcpp/impl/codegen/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
         "grpcpp/impl/codegen/server_context.h",
