+++ test/cpp/codegen/compiler_test_golden
@@ -36,7 +36,7 @@
 #include <grpcpp/support/message_allocator.h>
 #include <grpcpp/support/method_handler.h>
 #include <grpcpp/impl/codegen/proto_utils.h>
-#include <grpcpp/impl/codegen/rpc_method.h>
+#include <grpcpp/impl/rpc_method.h>
 #include <grpcpp/impl/codegen/server_callback.h>
 #include <grpcpp/impl/codegen/server_callback_handlers.h>
 #include <grpcpp/impl/codegen/server_context.h>
